refactor: add handler for clearing the cart
